use web worker for file hashing
